refactor: extract background jobs crate
