+++ Cargo.toml
@@ -23,10 +23,10 @@ publish = false
 
 [features]
 default = []
-enterprise = ["openobserve-api/enterprise", "openobserve-core/enterprise"]
-vectorscan = ["openobserve-api/vectorscan", "openobserve-core/vectorscan"]
-cloud = ["openobserve-api/cloud", "openobserve-core/cloud"]
-vortex = ["config/vortex", "openobserve-api/vortex", "openobserve-core/vortex"]
+enterprise = ["openobserve-api/enterprise", "openobserve-core/enterprise", "openobserve-jobs/enterprise"]
+vectorscan = ["openobserve-api/vectorscan", "openobserve-core/vectorscan", "openobserve-jobs/vectorscan"]
+cloud = ["openobserve-api/cloud", "openobserve-core/cloud", "openobserve-jobs/cloud"]
+vortex = ["config/vortex", "openobserve-api/vortex", "openobserve-core/vortex", "openobserve-jobs/vortex"]
 mimalloc = ["dep:mimalloc"]
 jemalloc = ["dep:tikv-jemallocator"]
 profiling = [
@@ -207,6 +207,7 @@ async-stream.workspace = true
 sourcemap.workspace = true
 openobserve-api.workspace = true
 openobserve-core.workspace = true
+openobserve-jobs.workspace = true
 
 [dev-dependencies]
 expect-test.workspace = true
@@ -219,6 +220,7 @@ rcgen.workspace = true
 members = [
     "src/api",
     "src/core",
+    "src/jobs",
     "src/config",
     "src/infra",
     "src/ingester",
@@ -246,6 +248,7 @@ flight = { path = "src/flight" }
 tantivy_utils = { path = "src/tantivy_utils" }
 openobserve-core = { path = "src/core" }
 openobserve-api = { path = "src/api" }
+openobserve-jobs = { path = "src/jobs" }
 o2_dex = { path = "src/enterprise/o2_dex" }
 o2_enterprise = { path = "src/enterprise/o2_enterprise", default-features = false }
 o2_openfga = { path = "src/enterprise/o2_openfga" }
