+++ src/api/src/handler/grpc/request/event.rs
@@ -96,7 +96,12 @@ impl Event for Eventer {
 
                 // Try batch download files
                 if !files_to_download.is_empty() {
-                    match crate::job::download_from_node(&grpc_addr, &files_to_download).await {
+                    match crate::service::file_downloader::download_from_node(
+                        &grpc_addr,
+                        &files_to_download,
+                    )
+                    .await
+                    {
                         Ok(failed) => failed_files = failed,
                         Err(e) => {
                             log::error!("[gRPC:Event] Failed to get files from notifier: {e}");
@@ -107,7 +112,7 @@ impl Event for Eventer {
 
                 // Fallback to individual downloads for failed files
                 for (id, account, file, size, ts) in failed_files {
-                    if let Err(e) = crate::job::queue_download(
+                    if let Err(e) = crate::service::file_downloader::queue_download(
                         TRACE_ID_FOR_CACHE_LATEST_FILE.to_string(),
                         id,
                         account,
@@ -124,7 +129,7 @@ impl Event for Eventer {
             } else {
                 // Direct download when download_from_node_enabled is false
                 for (id, account, file, size, ts) in files_to_download {
-                    if let Err(e) = crate::job::queue_download(
+                    if let Err(e) = crate::service::file_downloader::queue_download(
                         TRACE_ID_FOR_CACHE_LATEST_FILE.to_string(),
                         id,
                         account,
