+++ src/core/src/lib.rs
@@ -20,7 +20,6 @@
 #[cfg(feature = "enterprise")]
 pub mod cipher;
 pub mod common;
-pub mod job;
 pub mod service;
 
 pub(crate) static USER_AGENT_REGEX_FILE: &[u8] = include_bytes!("../ua_regex/regexes.yaml");
