+++ src/core/src/service/alerts/scheduler/handlers.rs
@@ -958,7 +958,7 @@ async fn handle_alert_triggers(
                     if let Some(batch) =
                         crate::service::alerts::grouping::get_ready_batch(&fingerprint)
                         && let Err(e) =
-                            crate::job::alert_grouping::send_grouped_notification_sync(batch).await
+                            crate::service::alerts::grouping::send_grouped_notification(batch).await
                     {
                         log::error!(
                             "[SCHEDULER trace_id {scheduler_trace_id}] Failed to send grouped notification: {}",
