+++ src/core/src/service/file_downloader.rs
