+++ src/core/src/service/mod.rs
@@ -28,6 +28,7 @@ pub mod dashboards;
 pub mod db;
 pub mod enrichment;
 pub mod enrichment_table;
+pub mod file_downloader;
 pub mod file_list;
 pub mod file_list_dump;
 pub mod folders;
