+++ src/core/src/service/self_reporting/mod.rs
@@ -53,11 +53,11 @@ mod triggers_schema;
 mod usage_schema;
 
 #[cfg(feature = "enterprise")]
-pub(crate) use evaluator_schema::ensure_evaluator_stream_initialized;
+pub use evaluator_schema::ensure_evaluator_stream_initialized;
 #[cfg(feature = "cloud")]
 pub use ingestion::ingest_data_retention_usages;
 #[cfg(feature = "enterprise")]
-pub(crate) use llm_scores_schema::ensure_llm_scores_stream_initialized;
+pub use llm_scores_schema::ensure_llm_scores_stream_initialized;
 
 pub async fn run() {
     #[cfg(not(feature = "enterprise"))]
