+++ src/jobs/src/job/promql_self_consume.rs
